// test/codegen/e2e/typescript/2020-12/embedded_resources/expected.d.ts
export type DocumentSystemItem = _DocumentSystemItem;

export type DocumentSystemAdditionalProperties = never;

export type DocumentSystemItemName = string;

export type DocumentSystemItemAdditionalProperties = never;

export interface _DocumentSystemItem {
  "name": DocumentSystemItemName;
}

export interface DocumentSystem {
  "item": DocumentSystemItem;
}